Implemented cross-wiki search.
* Use Lucene's MultiSearcher rather than merging indices for a few reasons:
 - It meant fewer changes to the indexing code.
 - Seprarte indices should be more robust.
 - It seems to be the way Lucene's designed to work.
 - Missing indices for foreign wikis need not be fatal.
* The configuration for each wiki now knows its own name, and a list of "other search index directories" where its searcher should search.
* Foreign wiki links use a new wiki attribute on the wikiPage tag rather than extending existing inter-wiki links. Inter-wiki links are semantically different, require manual setup, and their indices aren't necessarily accessible.
* The Lucene index stores a new wiki field (obviously) and a new uid field, since path can no longer be the key. a single uidFor(wikiName, pageName) method encapsulates the format of the uid which is currently wikiName::pageName.
* Results are grouped by wiki, with the current wiki first. For a given wiki, results are returned in rank order as before.

Things that might be worthy of a review:

the page(...) methods are a bit awkward now. We would want to overload to make the wiki name optional, but because of the parameter types we can't.

The way a new <ul> is started to break up the search results is efficient, but perhaps a bit intricate for JSP. It was done this way to avoid having to change the format of ${results}.