Switch the for-loops to for-each style.
Switch a use of StringBuffer to StringBuilder as the use appears thread-safe.

Switch a use of StringBuilder with no initial capacity to one with an initial capacity.

More contraversial: switch a massive if-(else-if...)-else block to a massive
switch (character) { } block. Contraversial because I'm really not sure whether
it is cleaner or not. No doubt mth will review :)