Add autosave for page edits using browser localstorage.
A small JS library sisyphus (originally in [1] with modifications in [2]) is used to autosave the contents of the editForm to browser local storage (save is triggered by any change to form data).

If there is stored data for the form when the user browses to the edit page then a 'restore' button is dispayed that restores the previous state of the form. This state is removed when the user clicks 'save'.

[1] https://github.com/simsalabim/sisyphus
[2] https://github.com/paulcadman/sisyphus

https://bugs.corefiling.com/show_bug.cgi?id=10370
supersedes: https://bugs.corefiling.com/show_bug.cgi?id=5871