Append the autocomplete widget to the searchBox in the page DOM.
The searchBox is now contained in a position:fixed container element. By
default the autocomplete widget is appended to the body of the page, and
so becomes disconnected from the searchBox when the page is scrolled. Using the autocomplete 'appendTo:' option we now append the widget to the searchBox so it remains attached when the page is scrolled.