Get rid of the RenderNode/ResultNode mess
Turns out that RenderNodes were how the regex renderer was doing a
"parse tree", and they were actually not necessary (probably) when using
a proper parser. I removed those and gave extra parameters to the
ResultNodes where necessary.

This brings the code significantly closer to being able to integrate
with the native testsuite.