Integrate with main project testsuite
This reveals two main classes of bugs:

 - Actual bugs
 - False positives due to different indentation/linebreaking of output.

Not sure how best to fix the latter class. The best solution is to
remove unimportant whitespace from the expected and actual results (but
that is work). I'll probably just ignore them, and rewrite the expected
results when the original renderer is replaced.