Speed up parsing significantly.
This first tries to parse the grammar as SLL(*) with no error recovery,
which is really fast if the page is valid. If an exception is thrown,
the default behaviour of LL(*) and lots of error recovery is tried,
which is much slower.

In practice, this worst-case fallback isn't a problem, as the grammar is
being constantly expanded to cope with de facto wikicreole use, reducing
errors all the time.