Correctly implement heading levels in DocbookRenderer.
The current section of interest is now the head of a stack which is
pushed to whenever a lower-level heading is encountered, and popped from
(and the old head appended to its containing section) when a
higher-level heading is encountered.

This has the knock-on effect of XSL-FO, rtf, PostScript, and PDF now
doing headings properly.