Upgrade SVNKit from version 1.3.2 to 1.3.8.
SVNKit 1.3.2 hardcodes the use of SSLv3 which is affected by the
POODLE exploit. SVNKit 1.3.6 and above will attempt to use the highest
available secure protocol to make connections. If this is not sufficient
then the 'svnkit.http.sslProtocols' Java system property may be used to
override the system-level available protocols to exclude the vunerable
SSLv3 protocol.

https://jira.int.corefiling.com/browse/REVIKI-570