The reviki lexer can now generate tokens for JIRA style links.
JIRA style links are surrounded by single square brackets and contain
the link title before the target:
[wikiword]
[title|http://target.com]

The lexer has this feature disabled by default. It can be enabled by
setting the jiraStyleLinks variable or calling the custom constructor.

Revikis parser has also been modified to deal with the new tokens.

The CreoleRenderer has a renderWithLexer method so which renderers may
use to configure the lexer.

https://jira.int.corefiling.com/browse/REVIKI-592