Prevent the terseblockquote rule from matching other blockquotes ahead in the text.
The terseblockquote rule matches arbitrary creole (as the markup in the
blockquote is rendered as creole) so this can match other blockquotes,
causing unintended nesting.

The solution used here is to have separate tokens for the start and end
of a terseblockquote in a way that just one blockquote can be matched at a
time.

NB: A lexer mode can't be used for this because the body of the quote
needs to be lexed with all rules from the default mode.

https://jira.int.corefiling.com/browse/REVIKI-611