Generate different AST nodes for nowiki, code and inline{code, nowiki}.
Nowiki markup in a line containing other markup is rendered as inline
nowiki. eg
"""
text {{{inline}}}
"""

And nowik markup on a single line is rendered as a block:
"""
{{{block}}}
"""

In contrast we wish to render single backtick delimited codeblocks inline, even if they're the only text on the line.

To do this the parser must generate different AST nodes for code and
nowiki blocks.

This commit also includes a change to the lexer so that single backtick
delimited codeblocks may have empty content.

https://jira.int.corefiling.com/browse/REVIKI-587