Fixes to the rendering of ``` code blocks to make it more like markdown.
The start of a code block must be of the form (modulo whitespace)
```<lanugage hint>?
<code>
```
and must appear at the start of a line (unless it occurs at the
beginning of a table cell).

Code block declarations that are not matched by the grammar are rendered
verbatim to help the user find errors in their markup.

So for example:
i) ```java some other text
ii) some other text ```java and other
both are rendered verbatim

The beginning of a code block is immediately followed by an EOF (so it
has no body) then it is not rendered at all.

https://jira.int.corefiling.com/browse/REVIKI-587