Render header, footer, sidebar ondemand when used by a template.
Adds ComplementaryPageRenderer that is added to each page as a request
atttribute. The methods on this class can be used as properties in the
JSP template. Each page is rendered only when the corresponding property
is accessed in the template. This avoids unnecessary rendering when a
tempalte does not use these pages.

Previously the header, footer and sidebar components were rendered and
added as attributes on the request for all pages. Some pages do not
display these components, for example:
* Edit pages
* Any page rendered with ctype=raw URL parameter
* Attachments

https://jira.int.corefiling.com/browse/REVIKI-642