Fix and test for https://jira.int.corefiling.com/browse/REVIKI-653.
The "login=force" code was a hang-over from a slightly different approach to the original authentication bug (https://jira.int.corefiling.com/browse/REVIKI-523) and isn't needed.  Testing for the non-existence of code is clearly hard, so I've tried to make a more general test the the URL given works.