Make ConfigPageCachingPageStore a ChangeSubscriber that expires cache entries in response to repository changes.
Previously a locked ConfigPage was always fetched from the external
repository which affected performance. Especially when
ConfigInterWikiLinks was locked.

The purpose of ConfigPageCachingPageStore.handleChanges is to expire
cached pages that have changed. When the cache is empty it does not
need to handle changes and so getHighestSyncedRevision() returns an
integer indicating to the change dispatcher that it doesn't need to be
sent changes.

https://jira.int.corefiling.com/browse/REVIKI-651