Add first-cut markdown support using commonmark.
commonmark makes this project require Java 6 (rather than Java 5) but
that was deemed OK.

There's now a per-page attribute "markdown" that switches the HTML
renderer between using the reviki parser/renderer and the markdown
parser/renderer. There's also a checkbox on the edit page to switch the
attribute on/off.